Move middleware installed global into agent
