+++ internal/agent/agent_test.go
@@ -2,41 +2,15 @@ package agent_test
 
 import (
 	"sync"
-	"sync/atomic"
 	"testing"
 	"time"
 
 	"github.com/AikidoSec/firewall-go/internal/agent"
 	"github.com/AikidoSec/firewall-go/internal/agent/aikido_types"
 	"github.com/AikidoSec/firewall-go/internal/agent/cloud"
-	"github.com/AikidoSec/firewall-go/internal/agent/globals"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestOnMiddlewareInstalled(t *testing.T) {
-	t.Run("sets MiddlewareInstalled to 1", func(t *testing.T) {
-		// Reset the value before test
-		atomic.StoreUint32(&globals.MiddlewareInstalled, 0)
-
-		agent.OnMiddlewareInstalled()
-
-		value := atomic.LoadUint32(&globals.MiddlewareInstalled)
-		assert.Equal(t, uint32(1), value, "MiddlewareInstalled should be set to 1")
-	})
-
-	t.Run("can be called multiple times", func(t *testing.T) {
-		// Reset the value before test
-		atomic.StoreUint32(&globals.MiddlewareInstalled, 0)
-
-		agent.OnMiddlewareInstalled()
-		agent.OnMiddlewareInstalled()
-		agent.OnMiddlewareInstalled()
-
-		value := atomic.LoadUint32(&globals.MiddlewareInstalled)
-		assert.Equal(t, uint32(1), value, "MiddlewareInstalled should remain 1 after multiple calls")
-	})
-}
 
 func TestOnDomain(t *testing.T) {
 	t.Run("calls storeDomain correctly", func(t *testing.T) {
